+++ src/main/js/ReadmeNavLink.tsx
@@ -1,10 +1,11 @@
 import React from "react";
 import { WithTranslation, withTranslation } from "react-i18next";
-import { NavLink } from "@scm-manager/ui-components";
+import { SecondaryNavigationItem } from "@scm-manager/ui-components";
 
 type Props = WithTranslation & {
   url: string;
   activeWhenMatch: (route: any) => boolean;
+  collapsed?: boolean;
 };
 
 class ReadmeNavLink extends React.Component<Props> {
@@ -12,11 +13,12 @@ class ReadmeNavLink extends React.Component<Props> {
     const { url, activeWhenMatch, t } = this.props;
 
     return (
-      <NavLink
+      <SecondaryNavigationItem
         to={`${url}/readme`}
         icon="fas fa-book-reader"
         label={t("scm-readme-plugin.navLink")}
         activeWhenMatch={activeWhenMatch}
+        title={t("scm-readme-plugin.navLink")}
       />
     );
   }