+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Add swagger rest annotations to generate openAPI specs for the scm-openapi-plugin. ([#2](https://github.com/scm-manager/scm-readme-plugin/pull/2))
+- Make navigation item collapsable ([#3](https://github.com/scm-manager/scm-readme-plugin/pull/3))
 
 ## 2.0.0-rc2 - 2020-01-29